organises all controllers and models into seperate files, also utilising routing in the app
