--- .env-example
@@ -1,1 +1,0 @@
-PGDATABASE=database_name_here
