+++ app.js
@@ -1,36 +1,7 @@
 const express = require("express");
 const app = express();
-const {
-  getTopics,
-  getAPI,
-  getArticle,
-  getArticles,
-  getComments,
-  postComment,
-  patchArticle,
-  deleteComment,
-  getUsers,
-} = require("./controllers/topics.controller.js");
-
-app.use(express.json());
-
-app.get("/api/topics", getTopics);
-
-app.get("/api", getAPI);
-
-app.get("/api/articles/:article_id", getArticle);
-
-app.get("/api/articles", getArticles);
-
-app.get("/api/articles/:article_id/comments", getComments);
-
-app.post("/api/articles/:article_id/comments", postComment);
-
-app.patch("/api/articles/:article_id", patchArticle);
-
-app.delete("/api/comments/:comment_id", deleteComment);
-
-app.get("/api/users", getUsers);
+const api = require("./routes/api-router.js");
+app.use("/api", api);
 
 app.all("*", (request, response) => {
   response.status(404).send({ msg: "Endpoint Not Found" });
