+++ controllers/articles.controllers.js
@@ -1,32 +1,11 @@
 const {
-  fetchTopics,
   fetchArticle,
   fetchArticles,
   fetchComments,
   checkForArticle,
   sendComment,
   updateArticle,
-  removeComment,
-  fetchUsers,
-} = require("../models/topics.models.js");
-fs = require("fs/promises");
-const path = `./endpoints.json`;
-
-function getTopics(req, res, next) {
-  return fetchTopics()
-    .then((topics) => res.status(200).send({ topics }))
-    .catch(next);
-}
-
-function getAPI(req, res, next) {
-  return fs
-    .readFile(path, "utf8")
-    .then((data) => {
-      let parsedData = { api: JSON.parse(data) };
-      res.status(200).send(parsedData);
-    })
-    .catch(next);
-}
+} = require("../models/");
 
 function getArticle(req, res, next) {
   const { article_id } = req.params;
@@ -73,29 +52,10 @@ function patchArticle(req, res, next) {
     .catch(next);
 }
 
-function deleteComment(req, res, next) {
-  const { comment_id } = req.params;
-  return removeComment(comment_id)
-    .then(() => {
-      res.status(204).send();
-    })
-    .catch(next);
-}
-
-function getUsers(req, res, next) {
-  return fetchUsers()
-    .then((users) => res.status(200).send({ users }))
-    .catch(next);
-}
-
 module.exports = {
-  getTopics,
-  getAPI,
   getArticle,
   getArticles,
   getComments,
   postComment,
   patchArticle,
-  deleteComment,
-  getUsers,
 };
