+++ models/comments.models.js
@@ -1,0 +1,13 @@
+function removeComment(comment_id) {
+  return db
+    .query("DELETE FROM comments WHERE comment_id = $1;", [comment_id])
+    .then(({ rowCount }) => {
+      if (rowCount === 0) {
+        return Promise.reject({ status: 404, msg: "comment does not exist" });
+      } else return;
+    });
+}
+
+module.exports = {
+  removeComment,
+};
