+++ models/users.models.js
@@ -1,0 +1,9 @@
+function fetchUsers() {
+  return db.query("SELECT * FROM users;").then(({ rows }) => {
+    return rows;
+  });
+}
+
+module.exports = {
+  fetchUsers,
+};
